backend: Changes
* Update JWT logic and its dependencies
* Add OAuth2 (Google) login dependencies and configs
* Update application.yml in order to grab params from env
* Add @ValidEmailAndPassword annotation for User entity

+++ src/main/java/it/unical/inf/ea/sefora_backend/dao/TokenDao.java
@@ -7,14 +7,14 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenDao extends JpaRepository<Token, Integer> {
+public interface TokenDao extends JpaRepository<Token, Long> {
 
     @Query(value = """
             select t from Token t inner join User u\s
             on t.user.id = u.id\s
             where u.id = :id and (t.expired = false or t.revoked = false)\s
             """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(Long id);
 
     Optional<Token> findByToken(String token);
 }
